Create a new class which holds information about input patterns
In order to remove many intermediate data structures which were
circulating between lexer generator passes, create a new PatternDesc
class which holds all necessary information about an input pattern.

Information consists of a pattern name, a block of code which should
execute if the pattern is recognized, an NFA transition matrix for this
pattern, as well as the list of DFA accepting states which recognize
this pattern.

        * regex_to_nfa.py (combine_nfas): Use PatternDesc class.
        * nfa_to_dfa.py (nfa_to_dfa): Ditto.
        * pattern_descriptor.py: New file.